improve npm generator

+++ packages/generator-reshow/generators/compile-sh/templates/compile.sh
@@ -6,7 +6,7 @@ DIR=$(
 )
 cd $DIR
 SWJS=${DIR}/service-worker.js
-webpackEnabled=$(awk -F "=" '/^webpackEnabled/ {print $2}' $DIR/.yo)
+webpackEnabled=$([ -e "$DIR/.yo" ] && awk -F "=" '/^webpackEnabled/ {print $2}' $DIR/.yo)
 
 conf='{'
 conf+='"assetsRoot":"./assets/",'
@@ -37,7 +37,7 @@ stop() {
     cat webpack.pid | xargs -I{} kill -9 {}
     npm run clean:webpack
   fi
-  rm $SWJS
+  [ -e "$SWJS" ] && rm $SWJS
   echo "Stop done"
 }
 
@@ -94,7 +94,7 @@ watch() {
 
 hot() {
   stop
-  rm $SWJS
+  [ -e "$SWJS" ] && rm $SWJS
   echo "Hot Mode"
   npm run build:es:ui -- --watch &
   npm run build:es:src -- --watch &
@@ -125,7 +125,7 @@ case "$1" in
     ;;
   *)
     develop
-    exit
+    exit 0
     ;;
 esac
